refactor(proxy): extract `get_ns_from_query` into dedicated query module
